+++ subprojects/mps-gradle-plugin/src/main/kotlin/com/specificlanguages/MpsDefaultsExtension.kt
@@ -3,6 +3,7 @@ package com.specificlanguages
 import org.gradle.api.file.DirectoryProperty
 import org.gradle.api.file.RegularFileProperty
 import org.gradle.api.provider.Property
+import org.gradle.jvm.toolchain.JavaLauncher
 import org.gradle.jvm.toolchain.JavaToolchainSpec
 
 /**
@@ -27,6 +28,15 @@ interface MpsDefaultsExtension {
 
     /**
      * Java executable that [RunAntScript] will use by default to run Ant.
+     *
+     * Prefer [javaLauncher] over [javaExecutable]. Setting [javaExecutable] overrides [javaLauncher].
      */
     val javaExecutable: RegularFileProperty
+
+    /**
+     * Java executable that [RunAntScript] will use by default to run Ant.
+     *
+     * Prefer [javaLauncher] over [javaExecutable]. Setting [javaExecutable] overrides [javaLauncher].
+     */
+    val javaLauncher: Property<JavaLauncher>
 }