+++ subprojects/mps-gradle-plugin/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.9.0 (unreleased)
+
+### Added
+
+- `mpsDefaults.javaLauncher` to specify the Java executable for MPS in a more modern way and support
+  the `com.specificlanguages.jbr-toolchain` plugin.
+
 ## 1.8.0
 
 ### Added